Release/1.6.0

+++ AirCasting/CoreData/MeasurementStreamStorage.swift
@@ -6,6 +6,7 @@ import CoreLocation
 import Foundation
 import Combine
 import SwiftUI
+import Resolver
 
 protocol MeasurementStreamStorage {
     func accessStorage(_ task: @escaping(HiddenCoreDataMeasurementStreamStorage) -> Void)
@@ -26,7 +27,7 @@ protocol MeasurementStreamStorageContextUpdate {
     func markStreamForDelete(_ sessionUUID: SessionUUID, sensorsName: [String], completion: () -> Void) throws
     func deleteSession(_ sessionUUID: SessionUUID) throws
     func deleteStreams(_ sessionUUID: SessionUUID) throws
-    func addNote(_ note: Note, for sessionUUID: SessionUUID) throws 
+    func addNote(_ note: Note, for sessionUUID: SessionUUID) throws
     func save() throws
 }
 
@@ -38,17 +39,11 @@ extension HiddenCoreDataMeasurementStreamStorage {
 
 final class CoreDataMeasurementStreamStorage: MeasurementStreamStorage {
 
-    private let persistenceController: PersistenceController
-    private lazy var updateSessionParamsService = UpdateSessionParamsService()
-    private let context: NSManagedObjectContext
-    let hiddenStorage: HiddenCoreDataMeasurementStreamStorage
-    
-    init(persistenceController: PersistenceController) {
-        self.persistenceController = persistenceController
-        self.context = persistenceController.editContext
-        self.hiddenStorage = HiddenCoreDataMeasurementStreamStorage(context: self.context)
-    }
-    
+    @Injected private var persistenceController: PersistenceController
+    @Injected private var updateSessionParamsService: UpdateSessionParamsService
+    private lazy var context: NSManagedObjectContext = persistenceController.editContext
+    private lazy var hiddenStorage = HiddenCoreDataMeasurementStreamStorage(context: self.context)
+
     /// All actions performed on CoreDataMeasurementStreamStorage must be performed
     /// within a block passed to this methood.
     /// This ensures thread-safety by dispatching all calls to the queue owned by the NSManagedObjectContext.
@@ -74,25 +69,25 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
         }
         completion()
     }
-    
+
     func markSessionForDelete(_ sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         sessionEntity.gotDeleted = true
         try context.save()
     }
-    
+
     enum Error: Swift.Error {
         case missingMeasurementStream
         case missingSensorName
     }
-    
-    private lazy var updateSessionParamsService = UpdateSessionParamsService()
+
+    @Injected private var updateSessionParamsService: UpdateSessionParamsService
     private let context: NSManagedObjectContext
-    
+
     init(context: NSManagedObjectContext) {
         self.context = context
     }
-    
+
     func deleteSession(_ sessionUUID: SessionUUID) throws {
         do {
             try context.delete(context.existingSession(uuid: sessionUUID))
@@ -100,7 +95,7 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
             Log.error("Error when deleting session")
         }
     }
-    
+
     func deleteStreams(_ sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         let toDelete = sessionEntity.allStreams!.filter({ $0.gotDeleted })
@@ -109,7 +104,7 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
         }
         forceUpdate(sessionEntity: sessionEntity)
     }
-    
+
     func forceUpdate(sessionEntity: SessionEntity) {
         sessionEntity.changesCount += 1
         // EXPLANATION for above line:
@@ -120,7 +115,7 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
     func save() throws {
         try self.context.save()
     }
-    
+
     func removeDuplicatedMeasurements(for sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         sessionEntity.allStreams?.forEach({ stream in
@@ -147,13 +142,13 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
         stream.addToMeasurements(newMeasurement)
 
         let session = stream.session
-        
+
         //otherwise dormant session status changes to active when syncing measurements
         if session?.status != .FINISHED {
             session?.status = .RECORDING
         }
     }
-    
+
     func addMeasurements(_ measurements: [Measurement], toStreamWithID id: MeasurementStreamLocalID) throws {
         let stream = try context.existingObject(with: id.id) as! MeasurementStreamEntity
 
@@ -165,7 +160,7 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
             stream.addToMeasurements(newMeasurement)
         }
     }
-    
+
     func saveThresholdFor(sensorName: String, thresholdVeryHigh: Int32, thresholdHigh: Int32, thresholdMedium: Int32, thresholdLow: Int32, thresholdVeryLow: Int32) throws {
         let existingThreshold: SensorThreshold? = try context.existingObject(sensorName: sensorName)
         if existingThreshold == nil {
@@ -184,26 +179,28 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
     }
 
     func createSessionAndMeasurementStream(_ session: Session, _ stream: MeasurementStream) throws -> MeasurementStreamLocalID {
-        let sessionEntity = SessionEntity(context: context)
+        let sessionEntity = newSessionEntity()
         updateSessionParamsService.updateSessionsParams(sessionEntity, session: session)
         return try saveMeasurementStream(for: sessionEntity, context: context, stream)
     }
-    
+
     func existingMeasurementStream(_ sessionUUID: SessionUUID, name: String) throws -> MeasurementStreamLocalID? {
         let session = try context.existingSession(uuid: sessionUUID)
         let stream = session.streamWith(sensorName: name)
         return stream?.localID
     }
-    
+
     func getExistingSession(with sessionUUID: SessionUUID) throws -> SessionEntity {
         let session = try context.existingSession(uuid: sessionUUID)
         return session
     }
-    
-    func updateMeasurements(stream: MeasurementStreamEntity, newMeasurements: NSOrderedSet) throws {
-        stream.measurements = newMeasurements
+
+    func removeAllMeasurements(in stream: MeasurementStreamEntity, except measurementsToLeave: [MeasurementEntity]) {
+        let idsToLeave = measurementsToLeave.map(\.objectID)
+        let measurementsToDelete = stream.allMeasurements?.filter { !idsToLeave.contains($0.objectID) } ?? []
+        measurementsToDelete.forEach { context.delete($0) }
     }
-    
+
     private func saveMeasurementStream(for session: SessionEntity, context: NSManagedObjectContext, _ stream: MeasurementStream) throws -> MeasurementStreamLocalID {
         let newStream = MeasurementStreamEntity(context: context)
         newStream.sensorName = stream.sensorName
@@ -233,14 +230,10 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
             threshold.thresholdHigh = stream.thresholdHigh
             threshold.thresholdVeryHigh = stream.thresholdVeryHigh
         }
-        // Save here is important so that NSManagedObjectID is not temporary.
-        try context.save()
         
-        try context.obtainPermanentIDs(for: [newStream])
-
         return newStream.localID
     }
-    
+
     func setStatusToFinishedAndUpdateEndTime(for sessionUUID: SessionUUID, endTime: Date?) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         sessionEntity.status = .FINISHED
@@ -253,7 +246,7 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
         sessionEntity.status = sessionStatus
         try context.save()
     }
-    
+
     func updateSessionEndtime(_ endTime: Date, for sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         sessionEntity.endTime = endTime.currentUTCTimeZoneDate
@@ -261,6 +254,11 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
         try context.save()
     }
     
+    func updateSessionEndTimeWithoutUTCConversion(_ endTime: Date, for sessionUUID: SessionUUID) throws {
+        let sessionEntity = try context.existingSession(uuid: sessionUUID)
+        sessionEntity.endTime = endTime
+    }
+
     func updateSessionNameAndTags(name: String, tags: String, for sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         sessionEntity.name = name
@@ -268,6 +266,11 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
         try context.save()
     }
     
+    func updateVersion(for sessionUUID: SessionUUID, to version: Int) throws {
+        let sessionEntity = try context.existingSession(uuid: sessionUUID)
+        sessionEntity.version = Int16(version)
+    }
+
     func updateSessionFollowing(_ sessionFollowing: SessionFollowing, for sessionUUID: SessionUUID) {
         do {
             let sessionEntity = try context.existingSession(uuid: sessionUUID)
@@ -275,13 +278,65 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
                 sessionEntity.followedAt = DateBuilder.getFakeUTCDate()
             } else {
                 sessionEntity.followedAt = nil
+                if let ui = sessionEntity.userInterface {
+                    context.delete(ui)
+                }
             }
             try context.save()
         } catch {
             Log.error("Error when saving changes in session: \(error.localizedDescription)")
         }
     }
+
+    func updateSessionOrder(_ order: Int, for sessionUUID: SessionUUID) {
+        do {
+            let sessionEntity = try context.existingSession(uuid: sessionUUID)
+            sessionEntity.rowOrder = Int64(order)
+            try context.save()
+        } catch {
+            Log.error("Error when saving changes in session: \(error.localizedDescription)")
+        }
+    }
+
+    func giveHighestOrder(to sessionUUID: SessionUUID) {
+        do {
+            let sessionEntity = try context.existingSession(uuid: sessionUUID)
+            let highestOrder = try context.getHighestRowOrder()
+            sessionEntity.rowOrder = (highestOrder ?? 0) + 1
+            try context.save()
+        } catch {
+            Log.error("Error when saving changes in session: \(error.localizedDescription)")
+        }
+    }
+
+    func setOrderToZero(for sessionUUID: SessionUUID) {
+        do {
+            let sessionEntity = try context.existingSession(uuid: sessionUUID)
+            sessionEntity.rowOrder = 0
+            try context.save()
+        } catch {
+            Log.error("Error when saving changes in session: \(error.localizedDescription)")
+        }
+    }
     
+    // MARK: - Notes
+
+    enum NoteStorageError: Swift.Error, LocalizedError {
+        case storageEmpty
+        case noteNotFound
+        case malformedStorageState
+        case multipleNotesFound
+
+        var errorDescription: String? {
+            switch self {
+            case .storageEmpty: return "Note storage is empty"
+            case .multipleNotesFound: return "Multiple notes for given ID found"
+            case .noteNotFound: return "No note with given ID found"
+            case .malformedStorageState: return "Data storage is in malformed state"
+            }
+        }
+    }
+
     func addNote(_ note: Note, for sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         let noteEntity = NoteEntity(context: context)
@@ -291,22 +346,22 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
         noteEntity.date = note.date
         noteEntity.number = Int64(note.number)
         sessionEntity.addToNotes(noteEntity)
-        try context.save()
     }
-    
+
     func updateNote(_ note: Note, newText: String, for sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
-        let note = (sessionEntity.notes?.first(where: { ($0 as! NoteEntity).number == note.number }) as! NoteEntity)
-        note.text = newText
-        try context.save()
+        if let note = (sessionEntity.notes?.first(where: { ($0 as! NoteEntity).number == note.number }) as? NoteEntity) {
+            note.text = newText
+        }
     }
-    
+
     func deleteNote(_ note: Note, for sessionUUID: SessionUUID) throws {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
-        let note = (sessionEntity.notes?.first(where: { ($0 as! NoteEntity).number == note.number }) as! NoteEntity)
-        context.delete(note)
+        if let note = (sessionEntity.notes?.first(where: { ($0 as! NoteEntity).number == note.number }) as? NoteEntity) {
+            context.delete(note)
+        }
     }
-    
+
     func getNotes(for sessionUUID: SessionUUID) throws -> [Note] {
         let sessionEntity = try context.existingSession(uuid: sessionUUID)
         return sessionEntity.notes?.map { note -> Note in
@@ -318,23 +373,38 @@ final class HiddenCoreDataMeasurementStreamStorage: MeasurementStreamStorageCont
                                    number: Int(n.number))
         } ?? []
     }
-    
+
     func fetchSpecifiedNote(for sessionUUID: SessionUUID, number: Int) throws -> Note {
         let session = try context.existingSession(uuid: sessionUUID)
-        let note = (session.notes?.first(where: { ($0 as! NoteEntity).number == number }) as! NoteEntity)
+        guard let allSessionNotes = session.notes else { throw NoteStorageError.storageEmpty }
+        let matching = try allSessionNotes.filter {
+            guard let note = $0 as? NoteEntity else { throw NoteStorageError.malformedStorageState }
+            return note.number == number
+        }
+        guard matching.count > 0 else { throw NoteStorageError.noteNotFound }
+        guard matching.count == 1 else { throw NoteStorageError.multipleNotesFound }
+        let note = matching[0] as! NoteEntity
+
         return Note(date: note.date ?? DateBuilder.getFakeUTCDate(),
                     text: note.text ?? "",
                     lat: note.lat,
                     long: note.long,
                     number: Int(note.number))
     }
+    
+    private func newSessionEntity() -> SessionEntity {
+        let sessionEntity = SessionEntity(context: context)
+        let uiState = UIStateEntity(context: context)
+        uiState.session = sessionEntity
+        return sessionEntity
+    }
 
     func createSession(_ session: Session) throws {
-        let sessionEntity = SessionEntity(context: context)
+        let sessionEntity = newSessionEntity()
         updateSessionParamsService.updateSessionsParams(sessionEntity, session: session)
         try context.save()
     }
-    
+
     func observerFor<T>(request: NSFetchRequest<T>) -> NSFetchedResultsController<T> {
         let frc = NSFetchedResultsController(fetchRequest: request,
                                              managedObjectContext: context,
@@ -349,21 +419,21 @@ final class PreviewMeasurementStreamStorage: MeasurementStreamStorage {
     func accessStorage(_ task: @escaping (HiddenCoreDataMeasurementStreamStorage) -> Void) {
         Log.info("accessing storage")
     }
-    
+
     func save() throws {
         Log.info("Faking saving ")
     }
-    
+
     func saveThresholdFor(sensorName: String, thresholdVeryHigh: Int32, thresholdHigh: Int32, thresholdMedium: Int32, thresholdLow: Int32, thresholdVeryLow: Int32) throws {
         Log.info("Faking saving thresholds")
     }
-    
+
     func updateSessionEndtime(_ endTime: Date, for sessionUUID: SessionUUID) throws {
         Log.info("Faking updating sessioon end time happened: \(endTime) for session \(sessionUUID)")
     }
-    
+
     func updateSessionFollowing(_ sessionStatus: SessionFollowing, for sessionUUID: SessionUUID) {}
-    
+
     func addMeasurement(_ measurement: Measurement, toStreamWithID id: MeasurementStreamLocalID) throws {
         Log.info("Nothing happened for \(measurement)")
     }
@@ -375,7 +445,7 @@ final class PreviewMeasurementStreamStorage: MeasurementStreamStorage {
     func createSessionAndMeasurementStream(_ session: Session, _ stream: MeasurementStream) throws -> MeasurementStreamLocalID {
         fatalError()
     }
-    
+
     func existingMeasurementStream(_ sessionUUID: SessionUUID, name: String) throws -> MeasurementStreamLocalID? {
         fatalError()
     }
